backend/requirements3.txt
certifi==2025.1.31
charset-normalizer==3.4.1
cramjam==2.9.1
fastparquet==2024.11.0
filelock==3.17.0
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
intersystems_irispython @ file:///Users/mgummow/Desktop/ICHack/hackathon-2024/install/intersystems_irispython-5.0.1-8026-cp38.cp39.cp310.cp311.cp312-cp38.cp39.cp310.cp311.cp312-macosx_10_9_universal2.whl#sha256=273268808c4375e18f6ea76e954e1de15aa6f7e79077561eccea8d8b5b0b3f1a
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
